requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
emailvalidator==0.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
GeoAlchemy2==0.12.5
geomet==0.3.0
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-decouple==3.6
requests==2.28.1
Shapely==1.8.5.post1
six==1.16.0
SQLAlchemy==1.4.43
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
